QWheelEvent not working on Cocoa 64Bit
The patch which added support for smooth scrolling will work only on 32
bit builds. The new methods (e.g. deviceDeltaX) of the NSEvent were
added with the float return type. This should be of CGFloat type
instead, which is different from float on 64 bit builds.

+++ src/gui/kernel/qcocoaview_mac.mm
@@ -64,9 +64,9 @@
 #include <qdebug.h>
 
 @interface NSEvent (DeviceDelta)
-  - (float)deviceDeltaX;
-  - (float)deviceDeltaY;
-  - (float)deviceDeltaZ;
+  - (CGFloat)deviceDeltaX;
+  - (CGFloat)deviceDeltaY;
+  - (CGFloat)deviceDeltaZ;
 @end
 
 QT_BEGIN_NAMESPACE
